requirements.txt
fastapi
flask
pymysql
python-dotenv
uvicorn
sqlalchemy
pydantic
requests
flake8
cryptography
bcrypt
python-jose